+++ src/main/java/uk/gov/hmcts/reform/em/annotation/service/dto/RectangleDTO.java
@@ -10,13 +10,13 @@ public class RectangleDTO extends AbstractAuditingDTO implements Serializable {
 
     private Long id;
 
-    private Float x;
+    private Double x;
 
-    private Float y;
+    private Double y;
 
-    private Float width;
+    private Double width;
 
-    private Float height;
+    private Double height;
 
     private Long annotationId;
 
@@ -28,35 +28,35 @@ public class RectangleDTO extends AbstractAuditingDTO implements Serializable {
         this.id = id;
     }
 
-    public Float getX() {
+    public Double getX() {
         return x;
     }
 
-    public void setX(Float x) {
+    public void setX(Double x) {
         this.x = x;
     }
 
-    public Float getY() {
+    public Double getY() {
         return y;
     }
 
-    public void setY(Float y) {
+    public void setY(Double y) {
         this.y = y;
     }
 
-    public Float getWidth() {
+    public Double getWidth() {
         return width;
     }
 
-    public void setWidth(Float width) {
+    public void setWidth(Double width) {
         this.width = width;
     }
 
-    public Float getHeight() {
+    public Double getHeight() {
         return height;
     }
 
-    public void setHeight(Float height) {
+    public void setHeight(Double height) {
         this.height = height;
     }
